requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
dateparser==1.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.22.2
pandas==1.4.1
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
websockets==9.1
yarl==1.7.2
